third_party/WebKit/LayoutTests/fast/dom/Window/orphaned-frame-access-expected.txt

This tests access to a window with a null frame. You should see "PASS" for each of the three tests below.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


missing property: PASS
missing array index: PASS
missing interface object: PASS
PASS successfullyParsed is true

TEST COMPLETE
